Make behavior of _tag_separator_filter consistent with spec
According to the spec for the separator function, it should check
against the model it is passed, not its own model.  And it should return
a boolean, so I guess on error it should default to returning False.

Note this doesn't seem to resolve the issue of it throwing and exception
on the bad itera.