Fix get_tasks() to use filters and drop get_task_ids() - LP: #534840
I think get_tasks() can just be an alias of get_tasks_filtered(['all'])
now.

I don't know that get_task_ids(status) is really useful enough on its
own.  I think this is basically redundant with get_task_ids_filtered()
which gives a more general purpose way to obtain a filtered list of task
ids.

Note the old code took a string of space-delimited status tags and
parsed them.  I think this can be done easily enough on the client-side,
and omitting it here avoids a little overhead.

The old code also supported passing status strings as unicode, and the
api would care for converting the string to 'ascii'.  I'm less certain
about this, but am likewise assuming this is something we can expect to
be done client-side and keep our logic simpler here.  In any case I was
getting an error on the normalize() call so if this is needed, the code
will need to be re-examined.  If support for inputting status (or filter
names) in unicode is important, maybe the filter names should be stored
as unicode all the way through, and not do the conversion to ascii at
all anyway.