Make `gtcli list` take space-separated filters/tags as criteria
Since the filters code accepts tags as filters, it is convenient to just
pass all the criteria directly in for filter arguments, and to make it
accept multiple filters, space-separated.  This enables `gtcli list` to
be useful for examining output for different filters and combinations of
filters.

(We leave a special case for 'today' as an alias for the workview,
although I wonder if this could either go away or become a filter
itself.)