The tests should not be disabled in trunk.
I've verified all the tests still pass fine on my system.  I guess they
were disabled simply for focusing on testing liblarch, but this is bad,
since it gives people false impressions that code committed to trunk is
okay, when it might not be.

Also, please do not disable tests without leaving some comment as to
*why* they are being disabled.