After giving it (even) more thoughts, I think it's better to set the due date to the most urgent parents due date when clearing it (instead of setting it to undefine as I did in the previous version of this patch).
Indeed, it reinforces the rule imposing that "task due date < parents
due dates" at all times. Since this the GTG policy, it's probably better
to insist on it (it requires to discover or learn the policy, but it
reduces the confusion once it is learned - a better doc would certainly
help here btw).

Moreover, in order to provide a better feedback on the fact that there
is a "default value" for a task due date depending on related task due
dates, I changed a "set due date" context menu label from "clear due
date" to "reset to default", since it could let one think that this
would actually unset the due date instead of setting it to its current
most constrained value (according to the due date policy).

All that being said, I think this actually corresponds to the previous
behavior, so this should not be a big deal...