Refactor code of Bugzilla plugin
There are two main changes. One is plugin's code is rewritten, however it
keeps original behavior to user. Another one is bugz is not a dependency,
because Bugzilla plugin just gets a bug information via Bugzilla's XMLRPC
Web service, and xmlrpclib is enough for this purpose. Furthermore, this
solves bug #1075703 naturally.